% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphicFunctions.R
\name{distripolygonHill}
\alias{distripolygonHill}
\title{One-side polygon from a distribution}
\usage{
distripolygonHill(distri, color, ysh = 0, maxdensity = NULL,
  transform = identity, adjust = 1)
}
\arguments{
\item{distri}{A vector containing the distribution to plot}

\item{color}{Color to fill the shape}

\item{ysh}{Shift on the y-axis}

\item{maxdensity}{NULL}

\item{logy}{Boolean}
}
\value{
None
}
\description{
Creates a polygon from a distribution
}
\examples{
distri <- rnorm(2000, 0,1)
plot(1, xlim=c(-3,3), type='n', ylim=c(-1,1))
distripolygonHill(distri=distri, color="red")

}
