% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphicFunctions.R
\name{PolyViolineSwitch}
\alias{PolyViolineSwitch}
\title{Switch between violine and one sided distribution}
\usage{
PolyViolineSwitch(violine = TRUE)
}
\arguments{
\item{violine}{Boolean, if TRUE (default) uses a violine distribution, otherwise, uses a one sided distribution}
}
\value{
distrifunction
}
\description{
Switches between violine and one sided distribution, for internal use
}
