\name{ViolineHills-package}
\alias{ViolineHills-package}
\alias{ViolineHills}
\docType{package}
\title{
\packageTitle{ViolineHills}
}
\description{
\packageDescription{ViolineHills}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ViolineHills}
\packageIndices{ViolineHills}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ViolineHills}

Maintainer: \packageMaintainer{ViolineHills}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
