% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphicFunctions.R
\name{distripolygonVioline}
\alias{distripolygonVioline}
\title{Illustration of crayon colors}
\usage{
distripolygonVioline(distri, color)
}
\arguments{
\item{method2order}{method to order colors (\code{"hsv"} or \code{"cluster"})}

\item{cex}{character expansion for the text}

\item{mar}{margin paramaters; vector of length 4 (see \code{\link[graphics]{par}})}
}
\value{
None
}
\description{
Creates a plot of the crayon colors in \code{\link{brocolors}}
}
\examples{
plot_crayons()

}
