% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphicFunctions.R
\name{plotdensities}
\alias{plotdensities}
\title{Plots multiple density distributions}
\usage{
plotdensities(distributions, violine = TRUE, rowleg = NULL,
  dlegend = NULL, globalmaxdensity = TRUE, sublegend = NULL,
  cexsubl = 1, sublegside = "left", btylegend = "o",
  blikefill = FALSE, legendncol = 1, transform = identity,
  adjust = 1, xshift = c(0, 0), yshift = c(0, 0), minden = 0.6,
  col = "rainbow", rowtext = NULL, ...)
}
\arguments{
\item{distributions}{A list, or list of lists containing the distributions to plot}

\item{violine}{Boolean. If TRUE draws a violine plot, otherwise, draws a one-sided distribution}

\item{dlegend}{Null or legend caption}

\item{globalmaxdensity}{Boolean}

\item{transform}{Function. Transform on the y-axis (e.g. log).}

\item{col}{Colors if specified Default is "rainbow" and creates colors based on a transparent raibow palette.}

\item{distrilayout}{Color to fill the shape}

\item{xdis}{X-label}
}
\value{
None
}
\description{
To plot multiple density distributions
}
\examples{
distri1 <- rnorm(2000, 0,1)
distri2 <- rnorm(2000, 1, 1)
plotdensities(distributions=list(distri1,distri2))

}
