% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphicFunctions.R
\name{plotdensities}
\alias{plotdensities}
\title{Plots multiple density distributions}
\usage{
plotdensities(distributions, distrilayout = NULL, xdis = "Value",
  violine = TRUE, dlegend = NULL, globalmaxdensity = TRUE,
  colgroups = NULL, legendncol = 1, transform = identity,
  adjust = 1, col = "rainbow")
}
\arguments{
\item{distributions}{A vector containing the distribution to plot}

\item{distrilayout}{Color to fill the shape}

\item{xdis}{X-label}

\item{violine}{Boolean. If TRUE draws a violine plot, otherwise, draws a one-sided distribution}

\item{dlegend}{Null or legend caption}

\item{globalmaxdensity}{Boolean}

\item{transform}{Function. Transform on the y-axis (e.g. log).}

\item{col}{Colors if specified Default is "rainbow" and creates colors based on a transparent raibow palette.}
}
\value{
None
}
\description{
To plot multiple density distributions
}
\examples{
distri1 <- rnorm(2000, 0,1)
distri2 <- rnorm(2000, 1, 1)
plotdensities(distributions=list(distri1,distri2))

}
